Move AboutTeam to hurumap/core
