+++ apps/climatemappedafrica/src/components/DataIndicators/DataIndicators.js
@@ -1,3 +1,4 @@
+import { Image } from "@hurumap/next";
 import { Box, ClickAwayListener, Grid, Slide } from "@mui/material";
 import PropTypes from "prop-types";
 import React, { useState } from "react";
@@ -6,7 +7,6 @@ import Icon from "./Icon";
 import IndicatorPanel from "./IndicatorPanel";
 
 import bg from "@/climatemappedafrica/assets/images/Mask Group 8.png";
-import Image from "@/climatemappedafrica/components/Image";
 import RichHeader from "@/climatemappedafrica/components/RichHeader";
 
 function DataIndicators({ indicators, title }) {
