+++ apps/climatemappedafrica/src/components/Hero/Hero.js
@@ -1,5 +1,6 @@
 import { Section } from "@commons-ui/core";
 import { RichTypography } from "@commons-ui/next";
+import { Image } from "@hurumap/next";
 import { Box, Grid, useMediaQuery } from "@mui/material";
 import dynamic from "next/dynamic";
 import PropTypes from "prop-types";
@@ -7,7 +8,6 @@ import React from "react";
 
 import heroBg from "@/climatemappedafrica/assets/images/bg-map-white.jpg";
 import DropdownSearch from "@/climatemappedafrica/components/DropdownSearch";
-import Image from "@/climatemappedafrica/components/Image";
 import RichHeader from "@/climatemappedafrica/components/RichHeader";
 
 const Map = dynamic(() => import("./Map"), { ssr: false });
