+++ packages/hurumap-core/package.json
@@ -60,16 +60,23 @@
   "peerDependencies": {
     "@babel/core": "catalog:",
     "@commons-ui/core": "workspace:*",
+    "@commons-ui/next": "workspace:*",
+    "@commons-ui/payload": "workspace:*",
     "@emotion/react": "catalog:",
     "@emotion/styled": "catalog:",
+    "@hurumap/next": "workspace:*",
     "@mui/icons-material": "catalog:",
     "@mui/material": "catalog:",
+    "@mui/utils": "catalog:",
+    "@mui/x-tree-view": "catalog:",
     "clsx": "catalog:",
     "deepmerge": "catalog:",
     "papaparse": "catalog:",
+    "prop-types": "catalog:",
     "react": "catalog:",
     "react-copy-to-clipboard": "catalog:",
     "react-dom": "catalog:",
+    "react-multi-carousel": "catalog:",
     "react-share": "catalog:",
     "vega": "catalog:",
     "xlsx": "catalog:"
