+++ packages/hurumap-core/src/AboutTeam/index.js
