+++ packages/hurumap-core/src/Card/Content.js
