+++ packages/hurumap-core/src/Carousel/Carousel.js
