+++ packages/hurumap-core/src/Carousel/Carousel.snap.js
