+++ packages/hurumap-core/src/Carousel/index.js
