+++ packages/hurumap-next/src/Image/Image.snap.js
