+++ packages/hurumap-next/src/index.js
@@ -1,3 +1,4 @@
 export { default as Source } from "./Source";
 export { default as Map } from "./Map";
 export { generateChoropleth } from "./Map/utils";
+export { default as Image } from "./Image";
